src/main/java/soon/PTCMR_Back/domain/team/service/TeamService.java
package soon.PTCMR_Back.domain.team.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import soon.PTCMR_Back.domain.team.entity.Team;
import soon.PTCMR_Back.domain.team.repository.TeamRepository;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class TeamService {

    private final TeamRepository teamRepository;


    public Long create(String title) {
        Team team = Team.create(title);

        Long result = teamRepository.save(team);

        return result;
    }
}
